Minor refactoring to minimize code duplication.
