+++ client/dashboard/sites/monitoring-http-responses-card/index.tsx
@@ -12,17 +12,15 @@ import {
 import { useViewportMatch } from '@wordpress/compose';
 import { useMemo, createElement, Element } from '@wordpress/element';
 import { Text } from '../../components/text';
+import {
+	convertTimeRangeToUnix,
+	chartColors,
+	getLineChartTickNumber,
+	getLineChartTickLabel,
+} from '../monitoring/utils';
 import MonitoringCard from '../monitoring-card';
 import type { HTTPCodeSerie } from './http-codes';
-import type { TimeRange } from '../monitoring/types';
 import type { Site, SiteHostingMetrics } from '@automattic/api-core';
-
-function convertTimeRangeToUnix( timeRange: number ): TimeRange {
-	const start = Math.floor( new Date().getTime() / 1000 ) - timeRange * 3600;
-	const end = Math.floor( new Date().getTime() / 1000 );
-
-	return { start, end };
-}
 
 type SiteMetricsData = {
 	responseStatusData: HttpStatusDataPoints | undefined;
@@ -101,7 +99,6 @@ function useSiteMetricsData(
 	};
 }
 
-const chartColors = [ '#3858E9', '#5BA300', '#F57600', '#B51963' ];
 const chartGlyphs = [
 	GlyphCircle,
 	GlyphCross,
@@ -176,39 +173,10 @@ export default function MonitoringHttpResponsesCard( {
 	}
 
 	const lessThanMediumViewport = useViewportMatch( 'medium', '<' );
-
-	let numTicks: undefined | number;
-	switch ( timeRange ) {
-		case 168:
-			numTicks = lessThanMediumViewport ? 3 : 7;
-			break;
-		case 72:
-			numTicks = lessThanMediumViewport ? 3 : 6;
-			break;
-		case 24:
-		case 6:
-			numTicks = lessThanMediumViewport ? 4 : 12;
-			break;
-	}
-
 	const xAxisOptions = {
-		tickFormat: ( date: string ) => {
-			const d = new Date( date );
-
-			if ( timeRange <= 24 ) {
-				return `${ d.getHours() }:${ d.getMinutes().toString().padStart( 2, '0' ) }`;
-			}
-
-			if ( timeRange > 72 || ( timeRange > 24 && lessThanMediumViewport ) ) {
-				return `${ d.toLocaleDateString() }`;
-			}
-
-			return `${ d.toLocaleDateString() } ${ d.getHours() }:${ d
-				.getMinutes()
-				.toString()
-				.padStart( 2, '0' ) }`;
-		},
-		numTicks: numTicks,
+		tickFormat: ( date: string ) =>
+			getLineChartTickLabel( date, timeRange, lessThanMediumViewport ),
+		numTicks: getLineChartTickNumber( timeRange, lessThanMediumViewport ),
 	};
 
 	const getLegendIcon = ( key: string ) => {
